requirements.txt
flask
numpy
opencv-python
torch
torchvision
pillow
ultralytics
pyyaml